+++ includes/class.llms.ajax.handler.php
@@ -2,7 +2,7 @@
 /**
  * LifterLMS AJAX Event Handler
  * @since    1.0.0
- * @version  [version]
+ * @version  3.14.9
  */
 if ( ! defined( 'ABSPATH' ) ) { exit; }
 
@@ -548,7 +548,7 @@ class LLMS_AJAX_Handler {
 	 * @param    [type]     $request  [description]
 	 * @return   [type]               [description]
 	 * @since    3.9.0
-	 * @version  [version]
+	 * @version  3.14.9
 	 */
 	public static function quiz_answer_question( $request ) {